### 🔥 Added
- `SfTable` component - reusable, sortable, filterable data table component

### 💪 Updated

- refactored `RfqsTable` to use `SfTable` component
