+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2021-07-24
+
+### 🔥 Added
+
+- `SfTable` component - reusable, sortable, filterable data table component
+
+### 💪 Updated
+
+- refactored `RfqsTable` to use `SfTable` component (instead of hardcoded logic)
+
 ## [1.3.0] - 2021-07-14
 
 ### 🔥 Added
