+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "riverdi-rfq--frontend",
-  "version": "1.3.0",
+  "version": "1.4.0",
   "description": "Frontend written in React Next.js, Typescript for Rivedi RFQ system",
   "private": true,
   "scripts": {
